+++ src/main/php/response/Cookie.php
@@ -173,17 +173,6 @@ class Cookie
     }
 
     /**
-     * returns name of cookie
-     *
-     * @return  string
-     * @deprecated since 4.0.0, use name() instead, will be removed with 5.0.0
-     */
-    public function getName()
-    {
-        return $this->name();
-    }
-
-    /**
      * returns value of cookie
      *
      * @return  string
@@ -191,17 +180,6 @@ class Cookie
     public function value()
     {
         return $this->value;
-    }
-
-    /**
-     * returns value of cookie
-     *
-     * @return  string
-     * @deprecated since 4.0.0, use value() instead, will be removed with 5.0.0
-     */
-    public function getValue()
-    {
-        return $this->value();
     }
 
     /**
@@ -215,17 +193,6 @@ class Cookie
     }
 
     /**
-     * returns expiration timestamp of cookie
-     *
-     * @return  int
-     * @deprecated since 4.0.0, use expiration() instead, will be removed with 5.0.0
-     */
-    public function getExpiration()
-    {
-        return $this->expiration();
-    }
-
-    /**
      * returns path of cookie
      *
      * @return  string
@@ -236,33 +203,11 @@ class Cookie
     }
 
     /**
-     * returns path of cookie
-     *
-     * @return  string
-     * @deprecated since 4.0.0, use path() instead, will be removed with 5.0.0
-     */
-    public function getPath()
-    {
-        return $this->path;
-    }
-
-    /**
      * returns domain of cookie
      *
      * @return  string
      */
     public function domain()
-    {
-        return $this->domain;
-    }
-
-    /**
-     * returns domain of cookie
-     *
-     * @return  string
-     * @deprecated since 4.0.0, use domain() instead, will be removed with 5.0.0
-     */
-    public function getDomain()
     {
         return $this->domain;
     }